Disable the forced enable Password Auth
Since Pre "split" of Vesta this was the default action when installing

Since 1.2.0 Hestia has support for addig / mananging of "RSA/SSH keys" via Web UI and Hestia has no issue with running like this.

As we don't change the default behaviour of the install there is no change of locking your self out.