Revise ssh / sshd service restarts
Debian / Ubuntu now alias sshd service to ssh service so restarting
ssh service is actually restarting it twice so only restart sshd
service. Furthermore rename ssh service restarts to sshd if it is the
only restart command. This works since sshd is the name of the service
on CentOS and Debian / Ubuntu now have aliases for sshd.