Admin's email in /etc/aliases
The purpose of the code is to place the admin's email address in /etc/aliases file, so that mails addressed to root will be forwarded to admin's email. To keep the code clean, I'm putting the comments here.
Line 1556: Make sure the file /etc/aliases does exist, so the installer doesn't throw any errors if it doesn't. Is it even possible that the file does not exist?
Line 1557: If there are other aliases for root, comment them out.
Line 1558: Before appending, make sure the file ends with a LF. Otherwise we'll be appending to the end of an existing line.
Line 1559: Do we also need to check if it's a valid email address? Is this being checked elsewhere?
Line 1564: If for some reason the file does not exist and we don't create it, then exim's config will fail to deliver messages on disk, filling-up paniclog.