Replace tabulation with \u0009
Tabulators in the value field will break the list of records. This probably should be made in a more global scale, as the json_decode function in my php server (PHP Version 7.4.22, running on Debian 10.10 4.19.0-17-amd64 #1 SMP Debian 4.19.194-2 (2021-06-21) x86_64 ) is returning an empty value if a value contains a tabulator (in my case a user introduced a TXT record that had a leading tab character).

I have read that the json spec wants tabulations represented by \t or \u0009.