Add --wait to iptables-restore
By default iptables exits with a failure with --wait=xxx xxx in seconds it wait for a max of xxx seconds before it returns an exit code. 

Adding a  waiting time for about 10 seconds does not affect any normal boot setup after lock has been obtained it will continue with the restore.