[Bug] Panel header (#2680)
* Fix issue with multiplying with text

When users set: Alias 10 / domain and unlimited domains an error is created. When domains is set to "Unlimited" and aliasses is set to a finite number "unlimted sign" number is shown. 

Also apply for mail accounts and dns records.
Ideally we need to change this behaviour

* Fix #2670 Display of path missing domain name

* Remove chkconfig

Deprecated by Ubuntu in 2013
https://forum.hestiacp.com/t/v-update-firewall-fail-in-ubuntu-20-04/6682

* #2681 Rebuild websites and mail domains after php version change

* #2678 DNS resolving fails on some new domains 

Allowing to test against Cloudflare to make sure it works will not take that longer after local look up fails