Add config to avoid restarting daemons on error
When calling any of the following scripts, the `send_email_report`
function will restart the service even though it knows that a
configuration error exists.

- `v-restart-cron`
- `v-restart-dns`
- `v-restart-ftp`
- `v-restart-proxy`
- `v-restart-web`

This patch adds a config variable, `NO_RESTART_ON_CONF_ERR` that must be
set to `'yes'` to avoid the restart behavior. This avoids making any
changes to the default behavior unless the user opts in.