Update v-update-firewall to replace correctly port in rules.conf and chains.conf files
Removed && [[ "$sshport" -ne "22" ]] in if because if port is 22 the port won't be changed in rules.conf and chains.conf files

Replaced sed -i "s/PORT='22' to allow any port in rules.conf file

Added sed command to replace the right ssh ports in chains.conf file. If we don't do this, those ssh ports won't be added to fail2ban-SSH chain.