Update v-list-sys-services to avoid 0 min uptime
Modify pidof command to use -d option to assign a delimiter so there is no need to use tr to change spaces by pipes.

In services with several pids, the script is taking the first pid in the list, but the first pid is the pid for last process, not the main process so a lot of times you can see an uptime of 0 mins instead of the actual uptime for the main/master process. So, I've modified it to get the last pid in the list instead of the first.