Auto delete spam and trash after 30 days
Users are used to spam and trash being cleaned up after some time
and for that reason don't manually cleanup this information. It's
also the go to way to get rid of credentials from mail that are
stored elsewere.

For that reason automatically removing spam and trash would enhance
user security but also server security since disks cannot fill up
by users not getting rid of their trash and spam.