Replaced "echo" with "sed" to avoid "Permission denied" in v-add-cron-letsencrypt-job
Replaced "echo" command with equivalent "sed" command as the "echo" resulted in "Permission denied". From what I can tell, this file is no longer used anywhere, but it has previously been updated to use the new "hestiaweb" user, so for completeness, I believe it should be updated to correctly update the crontab file as well, instead of erroring with "Permission denied".