Ensure composer alias uses the user-defined PHP CLI (#5155)
This change updates the creation of the composer alias so that it explicitly runs `composer` using the PHP CLI defined for the user rather than the system default. It also removes any previous composer alias to prevent conflicts and ensures consistent behavior with the user-specific PHP configuration.

Forum post: https://forum.hestiacp.com/t/composer-uses-wrong-php-version-but-php-v-shows-the-correct-one/20593